+++ lib/functions/vSysPort.ts
@@ -8,8 +8,8 @@
  *
  * @returns {boolean} Whether or not it's valid
  */
-function vSysPort (val:number):boolean {
-    return Number.isInteger(val) && val > 0 && val <= 65535;
+function vSysPort (val:unknown):val is number {
+    return Number.isInteger(val) && (val as number) > 0 && (val as number) <= 65535;
 }
 
 export {vSysPort, vSysPort as default};